fix sankey and use dynamic data
